solution at live 6

+++ src/main/java/hu/webuni/hr/minta/service/HolidayRequestService.java
@@ -10,6 +10,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
@@ -17,6 +19,7 @@ import org.springframework.util.StringUtils;
 import hu.webuni.hr.minta.dto.HolidayRequestFilterDto;
 import hu.webuni.hr.minta.model.Employee;
 import hu.webuni.hr.minta.model.HolidayRequest;
+import hu.webuni.hr.minta.model.HrUser;
 import hu.webuni.hr.minta.repository.HolidayRequestRepository;
 
 
@@ -28,6 +31,7 @@ public class HolidayRequestService {
 
 	@Autowired
 	EmployeeService employeeService;
+	
 
 	public List<HolidayRequest> findAll() {
 		return holidayRequestRepository.findAll();
@@ -74,8 +78,12 @@ public class HolidayRequestService {
 	}
 
 	@Transactional
-	public HolidayRequest approveHolidayRequest(long id, long approverId, boolean status) {
+	public HolidayRequest approveHolidayRequest(long id,  boolean status) {
 		HolidayRequest holidayRequest = holidayRequestRepository.findById(id).get();
+		long approverId = getCurrentUser().getEmployee().getEmployeeId();
+		Employee employee = holidayRequest.getEmployee();
+		if(employee.getManager().getEmployeeId() != approverId)
+			throw new AccessDeniedException("HolidayRequest to be approved is not onwed by a managed employee  of the current user");
 		holidayRequest.setApprover(employeeService.findById(approverId).get());
 		holidayRequest.setApproved(status);
 		holidayRequest.setApprovedAt(LocalDateTime.now());
@@ -98,8 +106,21 @@ public class HolidayRequestService {
 		HolidayRequest holidayRequest = holidayRequestRepository.findById(id).get();
 		if (holidayRequest.getApproved() != null)
 			throw new InvalidParameterException();
+		
+		checkOwnerOfHolidayRequest(holidayRequest);
+		
 		holidayRequest.getEmployee().getHolidayRequests().remove(holidayRequest);
 		holidayRequestRepository.deleteById(id);
+	}
+
+	private void checkOwnerOfHolidayRequest(HolidayRequest holidayRequest) {
+		HrUser hrUser = getCurrentUser();
+		if(!holidayRequest.getEmployee().getEmployeeId().equals(hrUser.getEmployee().getEmployeeId()))
+			throw new AccessDeniedException("HolidayRequest is not onwed by current user");
+	}
+
+	private HrUser getCurrentUser() {
+		return (HrUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 	}
 
 }
